+++ src/Translators/BaiduTranslate/BaiduTranslateEndpoint.cs
@@ -36,8 +36,8 @@ namespace BaiduTranslate
             case "zh-Hans":
             case "zh-CN":
                return "zh";
-            case "ja":
-               return "jp";
+            //case "ja":
+            //   return "jp";
             default:
                return lang;
          }