+++ src/XUnity.Common/Utilities/FastReflectionDelegate.cs
@@ -7,5 +7,11 @@ using System.Text;
 
 namespace XUnity.Common.Utilities
 {
+   /// <summary>
+   /// WARNING: Pubternal API (internal). Do not use. May change during any update.
+   /// </summary>
+   /// <param name="target"></param>
+   /// <param name="args"></param>
+   /// <returns></returns>
    public delegate object FastReflectionDelegate( object target, params object[] args );
 }