+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ﻿### 4.11.2
  * MISC - Add toggle to enable/disable translation aggregator UI
  * MISC - Enable copy translations / original text to clipboard from translation aggregator UI
+ * MISC - Enable "auto" source language for google, bing and papago translate
  * BUG FIX - Fixed some startup errors that occurred while initializing the plugin if the Resource Redirector was not present
- * BUG FIX - Fixed a bug in BaiduTranslate to allow it to use japanese (ja) as input/output language
 
 ### 4.11.1
  * MISC - Added support for named and optional groups in splitter and standard regexes